Finished challenge 4.21

+++ Web Development/Full-Stack-Open/part4 Testing-Express-and-user-auth/blog-list/controllers/blogs.js
@@ -52,6 +52,13 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
+
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  
   try {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
